Remove 'return' in global scope
